Rain Prediction
In this project whether it rains tomorrow or not is predicted by using a "Rain in Australia" kaggle dataset. A csv file is generated to compare with original dataset if the predictions are correct or not. Also, in this project the model is checked for accuracy. The algorithm used in this project is Random Forest.